Adds rudimentary ScanJavaLine ScanFn
ScanJavaLine is responsible for scanning essentially the first "line" of
Java code directly following a Javadoc. This can really be multiple
actual lines-- basically it scans everything up to the first curly
brace.

Currently, it does not differentiate between Java types and identifiers,
but the parser can use context to figure out which one is which (given
we're only parsing class and function definitions essentially)