Define a visitor interface, and implement the MarkdownVisitor.
A visitor object is one which implements the visit() function. This
function is passed a Document object, and performs custom functionality
depending on the visitor.

In addition to a MarkdownVisitor, the intention is to also build a
ValidationVisitor and SymbolVisitor, both of which will inspect and
provide more context for the MarkdownVisitor.