Add support for scanning parameters, and refactor Block.Text.
Prior to this commit, Block.Text was a plain string. However, there are
certain pieces of Javadoc syntax (namely params) which can be
interspersed anywhere in blocks of documentation text. This change
refactors Block.Text from a type of string -> []parser.Token.

Next we'll have to fix `MarkdownVisitor` to take the new Block.Text into
account.