Work towards handling JSX tags.
We need to detect embedded html tags within javadoc comments, since
there are some differences with how Docusaurus handles them.
Technically, tags in docusaurus are interpreted as JSX tags, not HTML,
and therefore doesn't offer affordance for things like self-closing
tags. This poses a problem since it is common in Javadocs to write
things like <p> or <br> without the closing '/'.

This commit adds support to the scanner and parser to ensure that
separate tokens for JSX tags get created. The next step will be to use a
stack in the parser to automatically close tags when unclosed tags are
found.